weechat: add netbsd_curses build option.
This build option is interesting because some usage conditions can
trigger hard-to-find ncurses bugs on musl and lead to a crash. The same
conditions don't appear to trigger crashes on weechat when built with
netbsd-curses.

The curses library being used doesn't appear in weechat's pkgconfig file
and therefore doesn't need to be in depends for weechat-devel.

Break subpackages list so the line isn't so long.